fix: adjusted menu behavior for better responsiveness and usability
